pyproject.toml
 [tool.poetry]
name = "hagent"
version = "0.1.2"
description = ""
authors = ["Farzaneh Rabiei <frabieik@ucsc.edu>" , "Jose Renau <renau@ucsc.edu>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
ruff = "^0.9.0"
pydoc-markdown = "^4.8.2"

[tool.ruff]
line-length = 130

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
litellm = "^1.59.10"
diskcache = "^5.6.3"
# legacy-cgi is needed for litlellm with python 3.13 (soon to be fixed, I hope)
legacy-cgi = "^2.6.2"
pyslang = "^8.0.0"
speechrecognition = "^3.14.1"
sounddevice = "^0.5.1"
numpy = "^2.2.2"
rapidfuzz = "^3.12.1"
ruamel-yaml = "^0.18.10"
openai = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["hagent"]
filterwarnings = [
    "ignore::DeprecationWarning",
   "ignore::UserWarning"
]

[tool.coverage.run]
omit = [
    "hagent/core/tests/*",
    "hagent/step/**/tests/*",
    "hagent/tool/**/tests/*",
]
